func_801D85D8/801E3918: added missing loop structure. Fixed from single-call to proper for-loop. Continuing structural refinement.

+++ src/kyoshin/CModelDisp.cpp
@@ -30,8 +30,8 @@ CModelDisp* func_801FBEB8(CModelDisp* self) {
 // Iterates 3 sub-objects and calls func_801FC2B4 with the mpController pointer.
 void func_801FC2B4(CModelDisp*, void**);
 void func_801FC0C4(CModelDisp* self) {
-    for (u8 i = 0; i < 3; i++) {
-        CModelDispSub* sub = (CModelDispSub*)((u8*)self + i * 0xFF0);
+    for (int i = 0; i < 3; i++) {
+        CModelDispSub* sub = (CModelDispSub*)((u8*)self + (u8)i * 0xFF0);
         func_801FC2B4(self, &sub->mpController);
     }
 }
